+ require github.com/89z/format v1.35.9

+++ details_test.go
@@ -57,7 +57,7 @@ func TestDetails(t *testing.T) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		head.AndroidID = uint64(device.AndroidID)
+		head.AndroidID = device.AndroidID
 		det, err := head.Details(app.id)
 		if err != nil {
 			t.Fatal(err)
